% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DiscreteSamples}
\alias{DiscreteSamples}
\title{DiscreteSamples}
\usage{
DiscreteSamples(skeletonList, n_samples, coordinate = -1L)
}
\arguments{
\item{skeletonList}{a piecewise deterministic skeleton (consisting of Times, Points and Velocities) returned by a sampler}

\item{n_samples}{number of samples to obtain}

\item{coordinate}{if specified, only obtain samples of the specified coordinate, otherwise obtain samples of all coordinates}
}
\value{
Returns a list containing the extracted samples and the times (on the continuous time scale) at which the samples are extracted
}
\description{
Extract discrete samples from a skeleton
}
