% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EstimateCovarianceMatrix}
\alias{EstimateCovarianceMatrix}
\title{EstimateCovarianceMatrix}
\usage{
EstimateCovarianceMatrix(skeletonList, coordinate = -1L,
  zeroMeans = FALSE)
}
\arguments{
\item{skeletonList}{a piecewise deterministic skeleton (consisting of Times, Points and Velocities) returned by a sampler}

\item{coordinate}{if specified, only estimate the variance of the specified coordinate, otherwise estimate the covariance matrix of all coordinates}

\item{zeroMeans}{if TRUE do not estimate means but assume a centered distribution}
}
\value{
Returns a list containing the estimated moment
}
\description{
Estimates the covariance matrix of a piecewise deterministic skeleton
}
