% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{EstimateMoment}
\alias{EstimateMoment}
\title{EstimateMoment}
\usage{
EstimateMoment(skeletonList, p, coordinate = -1L)
}
\arguments{
\item{skeletonList}{a piecewise deterministic skeleton (consisting of Times, Points and Velocities) returned by a sampler}

\item{p}{moment to estimate}

\item{coordinate}{if specified, only estimate the ESS of the specified coordinate, otherwise estimate the ESS of all coordinates}
}
\value{
Returns a list containing the estimated moment
}
\description{
Estimates the p-th moment of a piecewise deterministic skeleton
}
